+++ packages/babel-plugin-jsx-sfc/src/utils/index.ts
@@ -3,3 +3,11 @@ export const SFC_LIB = 'jsx-sfc';
 export const SFC_FUNC = 'sfc';
 
 export const SFC_COMPONENT = 'Component';
+
+export const SFC_CREATE_FUNC_RESULTS = 'createFuncResults';
+
+export const SFC_FORWARD_REF = 'forwardRef';
+
+export function getFuncResultsName(line: number) {
+  return `$sfcFuncResults_${line}`;
+}